% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARCH.R
\name{ARCH}
\alias{ARCH}
\title{ARCH}
\usage{
ARCH(x)
}
\arguments{
\item{x}{Rope}
}
\value{
value
}
\description{
Rope reduction to elliptical arch.
}
\examples{
# Test data
ROPE <- 50000

value <- ARCH(ROPE)
print(value)
}
