% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cartesian.R
\name{cartesian}
\alias{cartesian}
\title{Cartesian}
\usage{
cartesian(a, b, c, d, digits = 4)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{b}{ELLIPSOIDAL HEIGHT.}

\item{c}{Radian longitude.}

\item{d}{Radian Latitude.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
To convert from Geodesic to Cartesian.
}
\examples{
# Longitude
g <- -71
m <- 18
s <- 44.86475

# Value in radians
rad_long <- radians(g,m,s)

# Latitude
g <- -33
m <- 38
s <- 30.123456

# Value in radians
rad_lat <- radians(g, m, s)
print(rad_lat)

# ELLIPSOIDAL HEIGHT (h)
h <- 31.885

# Ellipsoids are: 1 = 'PSAD-56', 2 = 'SAD-69',	3 = 'WGS-84',	4 ='GRS-80 (SIRGAS)'.
value <- cartesian(4, h, rad_long, rad_lat, digits = 4)
print(value)
}
