% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E2.R
\name{E2}
\alias{E2}
\title{E2}
\usage{
E2(x)
}
\arguments{
\item{x}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}
}
\value{
value
}
\description{
To calculate 1-e^2
}
\examples{
# Ellipsoids are: 1 = 'PSAD-56', 2 = 'SAD-69',	3 = 'WGS-84',	4 ='GRS-80 (SIRGAS)'.
value <- E2(4)
print(value)
}
