% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geodis.R
\name{geodis}
\alias{geodis}
\title{Geodis}
\usage{
geodis(x, y, digits = 4)
}
\arguments{
\item{x}{Ellipsoidal height.}

\item{y}{Horizontal distance.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
Reduction of the horizontal distance to the Ellipsoid.
}
\examples{
# Test data
h <- 2500
Dhz <- 728.5

value <- geodis(h, Dhz, digits = 4)
print(value)
}
