% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lat_lon_height.R
\name{Lat_lon_height}
\alias{Lat_lon_height}
\title{Lat_lon_height}
\usage{
Lat_lon_height(a, x, y, z)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{x}{Cartesian x value.}

\item{y}{Cartesian y value.}

\item{z}{Cartesian z value.}
}
\value{
value
}
\description{
To convert from Cartesian to Geodesic.
}
\examples{
# Cartesian data
x <- 1711591.78090565
y <- -5060304.1659587
z <- -3473256.69328603

# Ellipsoids are: 1 = 'PSAD-56', 2 = 'SAD-69',	3 = 'WGS-84',	4 ='GRS-80 (SIRGAS)'.
value <- Lat_lon_height(4,x,y,z)
print(value)
}
