% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongLatToUTM.R
\name{LongLatToUTM}
\alias{LongLatToUTM}
\title{LongLatToUTM}
\usage{
LongLatToUTM(x, y, units = "m")
}
\arguments{
\item{x}{Sexagesimal longitude.}

\item{y}{Sexagesimal latitude.}

\item{units}{Select units for UTM to work.}
}
\value{
data.frame
}
\description{
To convert from Longitude and Latitude to UTM
}
\examples{
# Lat
g <- -33
m <- 12
s <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g, m, s)

# Lon
g1 <- -71
m1 <- 18
s1 <- 44.86475

# Value in sexagesimal
sexa_lon <- sexagesimal(g1, m1, s1)

value <- LongLatToUTM(sexa_lon, sexa_lat, units = 'm')
print(value)
}
