% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.R
\name{M}
\alias{M}
\title{M}
\usage{
M(x, y)
}
\arguments{
\item{x}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{y}{Latitude in radians}
}
\value{
value
}
\description{
To calculate the value of M
}
\examples{
# Lat
g <- -33
m <- 38
s <- 30.123456

value <- radians(g, m, s)
print(value)

# Ellipsoids are: 1 = 'PSAD-56', 2 = 'SAD-69',	3 = 'WGS-84',	4 ='GRS-80 (SIRGAS)'.
value1 <- M(4, value)
print(value1)
}
