% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radians.R
\name{radians}
\alias{radians}
\title{Radians}
\usage{
radians(x, y, z)
}
\arguments{
\item{x}{Value of Degree in Latitude or Longitude.}

\item{y}{Value of Minute in Latitude or Longitude.}

\item{z}{Value of Seconds in Latitude or Longitude.}
}
\value{
value
}
\description{
Convert separated data in Degrees Minutes and Seconds to Radians.
}
\examples{
# Lat
g <- -33
m <- 38
s <- 30.123456
value <- radians(g, m, s)
print(value)
}
