% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{read_data}
\alias{read_data}
\title{Read data in csv.}
\usage{
read_data(x, digits = 4)
}
\arguments{
\item{x}{the name of the file which the data are to be read from. this is a \code{data.frame} in format csv.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
a \code{tibble}
}
\description{
Read data in csv.
}
\examples{
\dontrun{
# Test data
csv <- system.file("extdata", "test.csv", package = "GeodesiCL")
data_test <- read_data(csv)
}
}
\seealso{
Package 'tibble' as 'readr'
}
