% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{SpatialPoints}
\alias{spTransform}
\alias{proj4string}
\alias{coordinates}
\alias{CRS}
\alias{mutate}
\alias{n_distinct}
\alias{mapview}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{n_distinct}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{mapview}{\code{\link[mapview:mapView]{mapview}}}

  \item{sp}{\code{\link[sp:CRS-class]{CRS}}, \code{\link[sp]{SpatialPoints}}, \code{\link[sp]{coordinates}}, \code{\link[sp:is.projected]{proj4string}}, \code{\link[sp]{spTransform}}}
}}

