% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scalfactor.R
\name{scalfactor}
\alias{scalfactor}
\title{Scalfactor}
\usage{
scalfactor(x, y, digits = 4)
}
\arguments{
\item{x}{East coordinate.}

\item{y}{Geodesic distance.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
Scale factor calculation.
}
\examples{
# Test data
EAST <- 224200
h <- 2500
Dhz <- 728.5

# Use the Geodesic distance from the function "Geodis"
arch_data <- geodis(h,Dhz)

value <- scalfactor(EAST, arch_data[2], digits = 4)
print(value)
}
