% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sin1.R
\docType{data}
\name{Sin_1}
\alias{Sin_1}
\title{Sin1}
\format{
A \code{numeric} with the value of sin 1"., which is:
\describe{
\item{Sin_1}{Value of sin 1".}
}
}
\usage{
Sin_1
}
\description{
Contains the values of sin 1".
}
\keyword{datasets}
