% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Read data for test}
\format{
A data frame with 19 columns:
\describe{
\item{Pt}{Contains the names of each point}
\item{East}{East UTM coordinate in a \code{tibble}}
\item{North}{North UTM coordinate in a \code{tibble}}
}
}
\usage{
test_data
}
\value{
a \code{tibble}
}
\description{
Read data for test
}
\examples{
# Test data
test_data
print(test_data)

# Test data option 2
date <- test_data
print(date)
}
\keyword{datasets}
