% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TO_TM.R
\name{TO_TM}
\alias{TO_TM}
\title{To convert from Geographic coordinate to TM.}
\usage{
TO_TM(a, longlat_df, d, e, f, g, digits = 4)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{longlat_df}{Sexagesimal longitude and latitude as dataframe.}

\item{d}{Central meridian.}

\item{e}{Scale factor Ko.}

\item{f}{False East (FE).}

\item{g}{False North (FN).}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
data.frame with the data in the following order: "East", "North", "X", "Y".
}
\description{
With this function it is possible to convert from Geographic coordinate to TM using the Central meridian, Scale factor Ko, False East, False North and obtain the decimal precision that you assign.
}
\examples{
# Test data
CM <- -69.00000
SC_FACTOR_Ko <- 0.99960
FE <- 500000.00000
FN <- 10000000.00000

# Point name
Pto <- "St1"

# Longitude
g <- -71
m <- 18
s <- 44.86475

# Value in sexagesimal
sexa_long <- sexagesimal(g, m, s)

# Latitude
g1 <- -33
m1 <- 12
s1 <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g1, m1, s1)

# Longitude and Latitude as data.frame
longlat_df <- as.data.frame(cbind(Pto,sexa_long,sexa_lat))

# ELLIPSOIDAL HEIGHT (h)
h <- 31.885

# Ellipsoids are: 1 = 'PSAD56', 2 = 'SAD69', 3 = 'WGS84', 4 = 'GRS80',
# 5 = 'GRS67', 6 = 'Airy 1830', 7 = 'Bessel 1841', 8 = 'Clarke 1880',
# 9 = 'Clarke 1866', 10 = 'International 1924', 11 = 'Krasovsky 1940'
value <- TO_TM(4, longlat_df, CM, SC_FACTOR_Ko, FE, FN, digits = 4)
print(value)
}
