% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_zone_hemisphere.R
\name{UTM_zone_hemisphere}
\alias{UTM_zone_hemisphere}
\title{UTM_zone_hemisphere}
\usage{
UTM_zone_hemisphere(x, y)
}
\arguments{
\item{x}{Sexagesimal longitude.}

\item{y}{Sexagesimal latitude.}
}
\value{
value
}
\description{
To find from Longitude and Latitude the UTM zone.
}
\examples{
# Lat
g <- -33
m <- 12
s <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g, m, s)

# Lon
g1 <- -71
m1 <- 18
s1 <- 44.86475

# Value in sexagesimal
sexa_lon <- sexagesimal(g1, m1, s1)

value <- UTM_zone_hemisphere(sexa_lon, sexa_lat)
print(value)
}
