% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMtoLongLat.R
\name{UTMtoLongLat}
\alias{UTMtoLongLat}
\title{UTMtoLongLat}
\usage{
UTMtoLongLat(utm_df, zone, hemisphere)
}
\arguments{
\item{utm_df}{East and North UTM in a data.frame.}

\item{zone}{Select UTM zone to work.}

\item{hemisphere}{select between north or south (written in lowercase).}
}
\value{
data.frame
}
\description{
To convert from UTM to Longitude and Latitude.
}
\examples{
# Point name
Pto <- "St1"

# Value for East
East <- 650012.58

# Value for North
North <- 5590735.41

# East and North as data.frame
utm_df <- as.data.frame(cbind(Pto,East,North))

# Zone
zone <- 18

# Hemisphere could be "north" or "south"
Hemisphere <- "south"

value <- UTMtoLongLat(utm_df,zone,Hemisphere)
print(value)
}
