% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch.R
\name{arch}
\alias{arch}
\title{Rope reduction to elliptical arch.}
\usage{
arch(x, digits = 4)
}
\arguments{
\item{x}{Rope}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
With this function it is possible to perform a reduction from chord to elliptical bow.
}
\examples{
# Test data
ROPE <- 50000

value <- arch(ROPE, digits = 4)
print(value)
}
