% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARCH.R
\name{ARCH}
\alias{ARCH}
\title{ARCH}
\usage{
ARCH(x, digits = 4)
}
\arguments{
\item{x}{Rope}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
Rope reduction to elliptical arch.
}
\examples{
# Test data
ROPE <- 50000

value <- ARCH(ROPE, digits = 4)
print(value)
}
