% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cartesian.R
\name{cartesian}
\alias{cartesian}
\title{Cartesian}
\usage{
cartesian(a, b, c, d)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{b}{ELLIPSOIDAL HEIGHT.}

\item{c}{Radian Latitude.}

\item{d}{Radian longitude.}
}
\value{
value
}
\description{
To convert from Geodesic to Cartesian.
}
\examples{
# Lat
g <- -33
m <- 38
s <- 30.123456

# Value in radians
rad_lat <- radians(g, m, s)
print(rad_lat)

# Lon
g <- -71
m <- 18
s <- 44.86475

# Value in radians
rad_lon <- radians(g,m,s)

# ELLIPSOIDAL HEIGHT (h)
h <- 31.885

# Ellipsoids are: 1 = 'PSAD-56', 2 = 'SAD-69',	3 = 'WGS-84',	4 ='GRS-80 (SIRGAS)'.
value <- cartesian(4,h, rad_lat, rad_lon)
print(value)
}
