% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E2.R
\name{E2}
\alias{E2}
\title{To calculate 1-e^2.}
\usage{
E2(x, digits = 4)
}
\arguments{
\item{x}{Selection of Ellipsoid.}

\item{digits}{Number of digits are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
To calculate the value for the function 1-e^2.
}
\note{
create data frame of epsg codes by epsg <- rgdal::make_EPSG()
}
\examples{
\donttest{
# To know the ellipsoids and the order open the Ellipsoids in the package
# and look for it number
Ellip <- Ellipsoids
#View(Ellip)

# We choose the number 5 which is GRS80
value <- E2(5, digits = 4)
print(value)
}
}
\references{
https://github.com/OSGeo/PROJ & https://github.com/cran/rgdal
}
