% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E3.R
\name{E3}
\alias{E3}
\title{To calculate 1-e^2*sen(lat)^2.}
\usage{
E3(x, y, digits = 4)
}
\arguments{
\item{x}{Selection of Ellipsoid.}

\item{y}{Latitude in radians.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
To calculate the value for the function 1-e^2*sen(lat)^2.
}
\note{
create data frame of epsg codes by epsg <- rgdal::make_EPSG()
}
\examples{
# Lat
g <- -33
m <- 38
s <- 30.123456

rad_lat <- radians(g, m, s)
print(rad_lat)

# To know the ellipsoids and the order open the Ellipsoids in the package and look for it number
Ellip <- Ellipsoids
#View(Ellip)

# We choose the number 5 which is GRS80
value <- E3(5, rad_lat, digits = 4)
print(value)
}
\references{
https://github.com/OSGeo/PROJ & https://github.com/cran/rgdal
}
