% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geodis.R
\name{geodis}
\alias{geodis}
\title{Geodis}
\usage{
geodis(x, y)
}
\arguments{
\item{x}{Ellipsoidal height.}

\item{y}{Horizontal distance.}
}
\value{
value
}
\description{
Reduction of the horizontal distance to the Ellipsoid.
}
\examples{
# Test data
h <- 2500
Dhz <- 728.5

value <- geodis(h,Dhz)
print(value)
}
