% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongLatToUTM.R
\name{LongLatToUTM}
\alias{LongLatToUTM}
\title{LongLatToUTM}
\usage{
LongLatToUTM(x, y, units = "m")
}
\arguments{
\item{x}{Sexagesimal longitude.}

\item{y}{Sexagesimal latitude.}

\item{units}{Select units for UTM to work.}
}
\value{
data.frame
}
\description{
To convert from Longitude and Latitude to UTM.
}
\examples{
# Longitude
g <- -71
m <- 18
s <- 44.86475

# Value in sexagesimal
sexa_long <- sexagesimal(g, m, s)

# Latitude
g1 <- -33
m1 <- 12
s1 <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g1, m1, s1)

value <- LongLatToUTM(sexa_long, sexa_lat, units = 'm')
print(value)
}
