% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongLatToUTM.R
\name{LongLatToUTM}
\alias{LongLatToUTM}
\title{LongLatToUTM}
\usage{
LongLatToUTM(longlat_df, units = "m", digits = 4)
}
\arguments{
\item{longlat_df}{Sexagesimal longitude and latitude as dataframe.}

\item{units}{Select units for UTM to work. DEFAULT: 'm'}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
data.frame
}
\description{
To convert from Longitude and Latitude to UTM.
}
\examples{
# Point name
Pto <- "St1"

# Longitude
g <- -71
m <- 18
s <- 44.86475

# Value in sexagesimal
sexa_long <- sexagesimal(g, m, s)

# Latitude
g1 <- -33
m1 <- 12
s1 <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g1, m1, s1)

# Longitude and Latitude as data.frame
longlat_df <- as.data.frame(cbind(Pto,sexa_long,sexa_lat))

value <- LongLatToUTM(longlat_df, units = 'm', digits = 4)
print(value)
}
