% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.R
\name{M}
\alias{M}
\title{To calculate the value of M.}
\usage{
M(x, y, digits = 4)
}
\arguments{
\item{x}{Selection of Ellipsoid to work between 1 = 'PSAD56', 2 = 'SAD69', 3 = 'WGS84', 4 = 'GRS80', 5 = 'GRS67', 6 = 'Airy 1830', 7 = 'Bessel 1841', 8 = 'Clarke 1880', 9 = 'Clarke 1866', 10 = 'International 1924', 11 = 'Krasovsky 1940'.}

\item{y}{Latitude in radians.}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
value
}
\description{
To calculate the value for the function of M.
}
\examples{
# Latitude
g <- -33
m <- 38
s <- 30.123456

rad_lat <- radians(g, m, s)
print(rad_lat)

# Ellipsoids are: 1 = 'PSAD56', 2 = 'SAD69', 3 = 'WGS84', 4 = 'GRS80',
# 5 = 'GRS67', 6 = 'Airy 1830', 7 = 'Bessel 1841', 8 = 'Clarke 1880',
# 9 = 'Clarke 1866', 10 = 'International 1924', 11 = 'Krasovsky 1940'
value <- M(4, rad_lat, digits = 4)
print(value)
}
