% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{SpatialPoints}
\alias{spTransform}
\alias{proj4string}
\alias{coordinates}
\alias{CRS}
\alias{mutate}
\alias{n_distinct}
\alias{mapview}
\alias{popupTable}
\alias{leaflet}
\alias{addTiles}
\alias{addMarkers}
\alias{HTML}
\alias{addProviderTiles}
\alias{as_tibble}
\alias{profvis}
\alias{read_csv}
\alias{show}
\alias{projInfo}
\alias{write.csv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{n_distinct}}}

  \item{htmltools}{\code{\link[htmltools]{HTML}}}

  \item{leaflet}{\code{\link[leaflet:map-layers]{addMarkers}}, \code{\link[leaflet]{addProviderTiles}}, \code{\link[leaflet:map-layers]{addTiles}}, \code{\link[leaflet]{leaflet}}}

  \item{leafpop}{\code{\link[leafpop:popup]{popupTable}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{mapview}{\code{\link[mapview:mapView]{mapview}}}

  \item{methods}{\code{\link[methods]{show}}}

  \item{profvis}{\code{\link[profvis]{profvis}}}

  \item{readr}{\code{\link[readr:read_delim]{read_csv}}}

  \item{rgdal}{\code{\link[rgdal]{projInfo}}}

  \item{sp}{\code{\link[sp:CRS-class]{CRS}}, \code{\link[sp]{SpatialPoints}}, \code{\link[sp]{coordinates}}, \code{\link[sp:is.projected]{proj4string}}, \code{\link[sp]{spTransform}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{utils}{\code{\link[utils:write.table]{write.csv}}}
}}

