% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scalfactor.R
\name{scalfactor}
\alias{scalfactor}
\title{Scalfactor}
\usage{
scalfactor(x, y)
}
\arguments{
\item{x}{East coordinate.}

\item{y}{Geodesic distance.}
}
\value{
value
}
\description{
Scale factor calculation.
}
\examples{
# Test data
EAST <- 224200

# Use the Geodesic distance from the function "Geodis"
arch_data <- geodis(h,Dhz)

value <- scalfactor(EAST, arch_data[2])
print(value)
}
