% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sexagesimal.R
\name{sexagesimal}
\alias{sexagesimal}
\title{Sexagesimal}
\usage{
sexagesimal(x, y, z)
}
\arguments{
\item{x}{Value of Degree in Latitude or Longitude.}

\item{y}{Value of Minute in Latitude or Longitude.}

\item{z}{Value of Seconds in Latitude or Longitude.}
}
\value{
value
}
\description{
Convert separated data in Degrees Minutes and Seconds to Decimal degrees.
}
\examples{
# Latitude
g <- -33
m <- 38
s <- 30.123456
value <- sexagesimal(g, m, s)
print(value)
}
