% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TO_TM.R
\name{TO_TM}
\alias{TO_TM}
\title{TO_TM}
\usage{
TO_TM(a, b, c, d, e, f, g)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{b}{Sexagesimal longitude.}

\item{c}{Sexagesimal latitude.}

\item{d}{Central meridian.}

\item{e}{Scale factor Ko.}

\item{f}{False East (FE).}

\item{g}{False North (FN).}
}
\value{
value
}
\description{
Geodesic coordinate transformation to TM.
}
\examples{
# Test data
CM <- -69.00000
SC_FACTOR_Ko <- 0.99960
FE <- 500000.00000
FN <- 10000000.00000

# Longitude
g <- -71
m <- 18
s <- 44.86475

# Value in sexagesimal
sexa_long <- sexagesimal(g, m, s)

# Latitude
g1 <- -33
m1 <- 12
s1 <- 27.11457

# Value in sexagesimal
sexa_lat <- sexagesimal(g1, m1, s1)

# ELLIPSOIDAL HEIGHT (h)
h <- 31.885

# Ellipsoids are: 1 = 'PSAD56', 2 = 'SAD69', 3 = 'WGS84', 4 = 'GRS80',
# 5 = 'GRS67', 6 = 'Airy 1830', 7 = 'Bessel 1841', 8 = 'Clarke 1880',
# 9 = 'Clarke 1866', 10 = 'International 1924', 11 = 'Krasovsky 1940'
value <- TO_TM(4, sexa_long, sexa_lat, CM, SC_FACTOR_Ko, FE, FN)
print(value)
}
