% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TO_TM.R
\name{TO_TM}
\alias{TO_TM}
\title{TO_TM}
\usage{
TO_TM(a, b, c, d, e, f, g)
}
\arguments{
\item{a}{Selection of Ellipsoid to work between 1 = 'PSAD-56', 2 = 'SAD-69',    3 = 'WGS-84',    4 ='GRS-80 (SIRGAS)'.}

\item{b}{Sexagesimal latitude.}

\item{c}{Sexagesimal longitude.}

\item{d}{Central meridian.}

\item{e}{Scale factor Ko.}

\item{f}{False East (FE).}

\item{g}{False North (FN).}
}
\value{
value
}
\description{
Geodesic coordinate transformation to TM
}
\examples{
# Test data
CM <- -69.00000
SC_FACTOR_Ko <- 0.99960
FE <- 500000.00000
FN <- 10000000.00000

# Lat
g <- -33
m <- 12
s <- 27.11457

# Value in radians
sexa_lat <- sexagesimal(g, m, s)

# Lon
g1 <- -71
m1 <- 18
s1 <- 44.86475

# Value in radians
sexa_lon <- sexagesimal(g1, m1, s1)

# ELLIPSOIDAL HEIGHT (h)
h <- 31.885

value <- TO_TM(4, sexa_lat, sexa_lon, CM, SC_FACTOR_Ko, FE, FN)
print(value)
}
