% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMtoLongLat.R
\name{UTMtoLongLat}
\alias{UTMtoLongLat}
\title{To convert from UTM to Geographic coordinate.}
\usage{
UTMtoLongLat(utm_df, zone, hemisphere = "south", digits = 4)
}
\arguments{
\item{utm_df}{Point name, East and North UTM in a data.frame.}

\item{zone}{Select UTM zone to work.}

\item{hemisphere}{select between north or south (written in lowercase). DEFAULT: "south"}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
a list with a data.frame [[1]] and leaflet map [[2]].
}
\description{
With this function it is possible to convert from UTM coordinate to Geographic coordinate.
}
\examples{
# Point name
Pto <- "St1"

# Value for East
East <- 650012.58

# Value for North
North <- 5590735.41

# East and North as data.frame
utm_df <- data.frame(Pto,East,North)

# Zone
zone <- 18

# Hemisphere could be "north" or "south"
hemisphere <- "south"

value <- UTMtoLongLat(utm_df, zone, hemisphere = "south", digits = 4)
print(value)
}
