% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMtoLongLat.R
\name{UTMtoLongLat}
\alias{UTMtoLongLat}
\title{To convert from UTM to Geographic coordinate.}
\usage{
UTMtoLongLat(utm_df, zone, hemisphere = "south", digits = 4)
}
\arguments{
\item{utm_df}{Point name, East and North UTM in a data.frame.}

\item{zone}{Select UTM zone to work.}

\item{hemisphere}{select between north or south (written in lowercase). DEFAULT: "south"}

\item{digits}{Number of digits the seconds are \code{\link{round}ed} to. DEFAULT: 4}
}
\value{
a list with a data.frame and leaflet map.
}
\description{
With this function it is possible to convert from UTM coordinate to Geographic coordinate.
}
\examples{
# Load test data from the package
csv <- system.file("extdata", "test.csv", package = "GeodesiCL")

# Point name, East and North as data.frame
utm_df <- data.frame(read_data(csv))

# Zone
zone <- 19

# Hemisphere could be "north" or "south"
hemisphere <- "south"

value <- UTMtoLongLat(utm_df, zone, hemisphere = "south", digits = 4)
print(value)
}
\seealso{
\code{data.frame}
}
